Move writer tracking into orchestrator module
